+++ R/bru.inference.R
@@ -94,6 +94,27 @@ bru_info_upgrade <- function(object,
       }
       object[["inlabru_version"]] <- "2.5.3.9003"
     }
+    if (utils::compareVersion("2.5.3.9005", old_ver) > 0) {
+      message("Upgrading bru_info to 2.5.3.9005")
+      # Make sure component$mapper is a bru_mapper_scale
+      for (k in seq_along(object[["model"]][["effects"]])) {
+        cmp <- object[["model"]][["effects"]][[k]]
+        # Convert offset mappers to const mappers
+        if (inherits(cmp$main$mapper, "bru_mapper_offset")) {
+          cmp$main$mapper <- bru_mapper_const()
+        }
+        cmp[["mapper"]] <-
+          bru_mapper_scale(
+            bru_mapper_multi(list(
+              main = cmp$main$mapper,
+              group = cmp$group$mapper,
+              replicate = cmp$replicate$mapper
+            ))
+          )
+        object[["model"]][["effects"]][[k]] <- cmp
+      }
+      object[["inlabru_version"]] <- "2.5.3.9005"
+    }
     object[["inlabru_version"]] <- new_version
   }
   object
@@ -1634,11 +1655,21 @@ lin_predictor <- function(lin, state) {
       function(x) {
         Ax_list <-
           lapply(names(x$A), function(xx) {
-            x[["A"]][[xx]] %*% state[[xx]]
+            if (is.null(state[[xx]])) {
+              stopifnot(is.null(x[["A"]][[xx]]) ||
+                          (ncol(x[["A"]][[xx]]) == 0))
+              NULL
+            } else {
+              x[["A"]][[xx]] %*% state[[xx]]
+            }
           })
         Ax <- do.call(cbind, Ax_list)
-        sumAx <- Matrix::rowSums(Ax)
-        as.vector(x$offset) + as.vector(sumAx)
+        if (!is.null(Ax)) {
+          sumAx <- Matrix::rowSums(Ax)
+          as.vector(x$offset) + as.vector(sumAx)
+        } else {
+          as.vector(x$offset)
+        }
       }
     )
   )