+++ tests/testthat/test_mapper.R
@@ -263,17 +263,17 @@ test_that("mapper collection automatic construction consistency", {
       ibm_values(cmp2$indep$mapper, inla_f = inla_f)
     )
     if (inla_f) {
-      input <- list(
+      input <- list(mapper = list(
         main = data$val,
         group = rep(1, 3),
         replicate = rep(1, 3)
-      )
+      ))
     } else {
-      input <- list(
+      input <- list(mapper = list(
         main = list(u = data$val),
         group = rep(1, 3),
         replicate = rep(1, 3)
-      )
+      ))
     }
     expect_identical(
       as.matrix(ibm_amatrix(cmp1$indep$mapper, input = input, inla_f = inla_f)),